Code cleanup and add anonymity toggle to community monitor settings page
